Added a $ssh param to $cmd.

+++ cli/src/lib/scriptExecutors/$cmd.ts
@@ -1,11 +1,11 @@
+/* eslint-disable no-template-curly-in-string */
+/* eslint-disable max-len */
 import child_process from 'child_process'
+import crypto from 'crypto'
 import fs from 'fs'
 import path from 'path'
-import { yellow } from '../colour.js'
-import crypto from 'crypto'
-import { isString, type ResolvedEnv } from '../types.js'
+import { isDefined, isDockerCmdScript, isSSHCmdScript, type CmdScript, type ResolvedEnv } from '../types.js'
 import { resolveResolveScript } from './ScriptExector.js'
-import logger from '../utils/logger.js'
 
 export const randomString = (): string => crypto.randomBytes(20).toString('hex')
 
@@ -23,6 +23,32 @@ export const cleanupOldTmpFiles = (env: ResolvedEnv): void => {
   }
 }
 
+const envToDockerEnvfile = (env: ResolvedEnv): string => {
+  // TODO make this configurable via ENVIRONMENT?
+  return Object.entries(env).map(([k, v]) => `${k}=${v}\n`).join('')
+}
+
+const envToShellExports = (env: ResolvedEnv): string => {
+  // TODO make this configurable via ENVIRONMENT?
+  return Object.entries(env).map(([k, v]) => `export ${k}="${v}"\n`).join('')
+}
+
+const writeShellScript = (filepath: string, content: string, env?: ResolvedEnv): void => {
+  // inject environment exports into content
+  if (isDefined(env)) {
+    const patt = /^#[^\n]+\n/gm
+    const envexports = envToShellExports(env)
+    const match = patt.exec(content)
+    if (match != null) {
+      content = content.substring(0, 28) + envexports + content.substring(28)
+    } else {
+      content = envexports + content
+    }
+  }
+  fs.writeFileSync(filepath, content, 'utf-8')
+  fs.chmodSync(filepath, 0o755)
+}
+
 /**
  * Executes the provided multiline command, and returns the stdout as a string.
  * @param multilineCommand
@@ -31,8 +57,7 @@ export const cleanupOldTmpFiles = (env: ResolvedEnv): void => {
  * @returns
  */
 export const executeCmd = (
-  multilineCommand: string,
-  dockerImage: string | undefined = undefined,
+  script: CmdScript,
   opts: any = undefined,
   env: ResolvedEnv
 ): string => {
@@ -42,29 +67,31 @@ export const executeCmd = (
     const filepath = path.resolve(`.tmp-${rand}.sh`)
     const envfile = path.resolve(`.env-${rand}.txt`)
     const parent = path.dirname(filepath)
-    const runInDocker = typeof dockerImage !== 'undefined'
-    fs.writeFileSync(filepath, multilineCommand, 'utf-8')
-    fs.chmodSync(filepath, 0o755)
-    if (runInDocker) {
-      fs.writeFileSync(envfile, Object.entries(opts.env).map(([k, v]) => `${k}=${v as string}`).join('\n'), 'utf-8')
-      fs.chmodSync(filepath, 0o644)
+
+    // run script based on underlying implementation
+    let output = ''
+    if (isDockerCmdScript(script)) {
+      // run on docker
+      writeShellScript(filepath, script.$cmd)
+      writeShellScript(envfile, envToDockerEnvfile(opts.env))
+      const DEFAULT_DOCKER_SCRIPT = 'docker run -t --rm --network host --entrypoint "" --env-file "${envfile}" -w "${parent}" -v "${parent}:${parent}" ${dockerImage} /bin/sh -c "chmod 755 ${filepath} && ${filepath}"'
+      const { DOCKER_SCRIPT: dockerScript = DEFAULT_DOCKER_SCRIPT } = env
+      const cmd = resolveResolveScript('-', { $resolve: dockerScript }, { envfile, filepath, dockerImage: script.$image, parent }, false)
+      output = child_process.execSync(cmd, opts)
+    } else if (isSSHCmdScript(script)) {
+      // run on remote machine
+      writeShellScript(filepath, script.$cmd, opts.env)
+      const DEFAULT_SSH_SCRIPT = 'ssh "${user_at_server}" < "${filepath}"'
+      const { SSH_SCRIPT: sshScript = DEFAULT_SSH_SCRIPT } = env
+      const cmd = resolveResolveScript('-', { $resolve: sshScript }, { envfile, filepath, user_at_server: script.$ssh, parent }, false)
+      output = child_process.execSync(cmd, opts)
+    } else {
+      // otherwise fallback to local machine
+      writeShellScript(filepath, script.$cmd, opts.env)
+      output = child_process.execSync(filepath, opts)
     }
-    // eslint-disable-next-line max-len, no-template-curly-in-string
-    const DEFAULT_DOCKER_SCRIPT = 'docker run -t --rm --network host --entrypoint "" --env-file "${envfile}" -w "${parent}" -v "${parent}:${parent}" ${dockerImage} /bin/sh -c "chmod 755 ${filepath} && ${filepath}"'
-    const { DOCKER_SCRIPT: dockerScript = DEFAULT_DOCKER_SCRIPT } = env
-    const cmd = runInDocker
-      ? resolveResolveScript('-', { $resolve: dockerScript }, { envfile, filepath, dockerImage, parent }, false)
-      : filepath
-    const output = child_process.execSync(cmd, opts)
-    // TODO - Remove if not required - program now deletes .env*.txt and .tmp*.sh files on startup
-    // if (runInDocker && fs.existsSync(envfile)) {
-    //   fs.unlinkSync(envfile)
-    // }
-    // if (fs.existsSync(filepath)) {
-    //   fs.unlinkSync(filepath)
-    // } else {
-    //   logger.warn(`warn: Could not delete ${filepath}`)
-    // }
+    // deliberately do not delete files on cleanup - so user has visibility for debugging.
+    // plus scripts are cleaned up on startup...
     return output !== null ? output.toString() : ''
   } catch (err: any) {
     const status = err.status as string
